// gestimowebfront-end/src/gs-api/src/models/appel-loyers-facture-dto.ts
/* tslint:disable */
export interface AppelLoyersFactureDto {
  id?: number;
  idAgence?: number;
  idCreateur?: number;
  periodeAppelLoyer?: string;
  statusAppelLoyer?: string;
  datePaiementPrevuAppelLoyer?: string;
  dateDebutMoisAppelLoyer?: string;
  dateFinMoisAppelLoyer?: string;
  periodeLettre?: string;
  moisUniquementLettre?: string;
  anneeAppelLoyer?: number;
  moisChiffreAppelLoyer?: number;
  descAppelLoyer?: string;
  montantLoyerBailLPeriode?: number;
  soldeAppelLoyer?: number;
  nomLocataire?: string;
  prenomLocataire?: string;
  genreLocataire?: string;
  emailLocatire?: string;
  idLocataire?: number;
  nomAgence?: string;
  telAgence?: string;
  compteContribuableAgence?: string;
  emailAgence?: string;
  mobileAgence?: string;
  regimeFiscaleAgence?: string;
  faxAgence?: string;
  sigleAgence?: string;
  bienImmobilierFullName?: string;
  abrvBienimmobilier?: string;
  nomPropietaire?: string;
  prenomPropietaire?: string;
  genrePropietaire?: string;
  mailProprietaire?: string;
  idBailLocation?: number;
  abrvCodeBail?: string;
  nouveauMontantLoyer?: number;
  ancienMontant?: number;
  pourcentageReduction?: number;
  messageReduction?: string;
  typePaiement?: string;
  solderAppelLoyer?: boolean;
  cloturer?: boolean;
  unLock?: boolean;
}